change to one add recipe fragment add footer navigation
