+++ app/src/main/AndroidManifest.xml
@@ -13,20 +13,6 @@
         android:theme="@style/Theme.HomeChef"
         tools:targetApi="31">
         <activity
-            android:name=".AddRecipeSecondActivity"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
-            android:name=".AddRecipeFirstActivity"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
