+++ app/src/main/java/com/example/homechef/MainActivity.java
@@ -4,12 +4,20 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+
 
 public class MainActivity extends AppCompatActivity {
+
+    BottomNavigationView bottomNavigationView;
+    AddRecipeFragment addRecipeFragment;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
+        bottomNavigationView = findViewById(R.id.navFooter);
+        getSupportFragmentManager().beginTransaction().replace(R.id.fragmentContainer, addRecipeFragment).commit();
     }
 }
