--- app/src/main/res/layout/activity_add_recipe_second.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".AddRecipeSecondActivity">
-
-    <Button
-        android:id="@+id/saveButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="שמור"
-        tools:layout_editor_absoluteX="258dp"
-        tools:layout_editor_absoluteY="565dp" />
-
-    <Button
-        android:id="@+id/cancelButton"
-        android:layout_width="wrap_content"
-        android:layout_height="48dp"
-        android:backgroundTint="@color/white"
-        android:shadowColor="@color/orange"
-        android:text="ביטול"
-        android:textColor="@color/orange"
-        android:textColorHighlight="@color/white"
-        app:strokeColor="@color/orange"
-        app:strokeWidth="1dp"
-        tools:layout_editor_absoluteX="64dp"
-        tools:layout_editor_absoluteY="561dp" />
-
-    <TextView
-        android:id="@+id/recipeTitle"
-        android:layout_width="316dp"
-        android:layout_height="35dp"
-        android:text="שם המתכון"
-        tools:layout_editor_absoluteX="45dp"
-        tools:layout_editor_absoluteY="102dp" />
-
-    <EditText
-        android:id="@+id/editTextTextMultiLine"
-        android:layout_width="323dp"
-        android:layout_height="385dp"
-        android:ems="10"
-        android:hint="הוראות המתכון שלי..."
-        android:gravity="start|top"
-        android:inputType="textMultiLine"
-        tools:layout_editor_absoluteX="44dp"
-        tools:layout_editor_absoluteY="156dp" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
