+++ app/src/main/res/layout/fragment_add_recipe.xml
@@ -1,0 +1,76 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".AddRecipeFragment">
+
+    <Button
+        android:id="@+id/uploadPicButton"
+        android:layout_width="200dp"
+        android:layout_height="40dp"
+        android:text="העלה תמונה"
+        app:cornerRadius="1dp"
+        tools:translationX="116dp"
+        tools:translationY="60dp"/>
+
+    <EditText
+        android:id="@+id/recipeTitle"
+        android:layout_width="293dp"
+        android:layout_height="wrap_content"
+        android:hint="כותרת מתכון"
+        android:inputType="text"
+        android:minHeight="48dp"
+        tools:translationX="74dp"
+        tools:translationY="100dp" />
+
+    <EditText
+        android:id="@+id/estimatedTime"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:ems="10"
+        android:hint="זמן הכנה משוער"
+        android:inputType="time"
+        android:minHeight="48dp"
+        tools:translationX="157dp"
+        tools:translationY="145dp" />
+
+    <Spinner
+        android:id="@+id/countryDropdown"
+        android:layout_width="208dp"
+        android:layout_height="38dp"
+        android:autofillHints="מאיזה מדינה המתכון שלך?"
+        tools:translationX="159dp"
+        tools:translationY="195dp" />
+
+    <EditText
+        android:id="@+id/editTextTextMultiLine"
+        android:layout_width="323dp"
+        android:layout_height="319dp"
+        android:ems="10"
+        android:gravity="start|top"
+        android:hint="הוראות המתכון שלי..."
+        android:inputType="textMultiLine"
+        tools:translationX="44dp"
+        tools:translationY="235dp" />
+
+    <Button
+        android:id="@+id/saveButton"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="שמור"
+        tools:translationX="240dp"
+        tools:translationY="620dp" />
+
+    <Button
+        android:id="@+id/cancelButton"
+        android:layout_width="wrap_content"
+        android:layout_height="48dp"
+        android:backgroundTint="@color/white"
+        android:text="ביטול"
+        android:textColor="@color/orange"
+        tools:translationX="85dp"
+        tools:translationY="620dp" />
+
+</FrameLayout>
