+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:id="@+id/profile"
+        android:title="הפרופיל שלי"
+        android:icon="@drawable/icon_person"/>
+
+    <item android:id="@+id/addRecipe"
+        android:title="הוסף מתכון"
+        android:icon="@drawable/icon_add_post"/>
+
+    <item android:id="@+id/home"
+    android:title="כל המתכונים"
+    android:icon="@drawable/icon_home"/>
+</menu>
